Suspected bug
I just suspect this to be a bug (due to a typo) as the argument `conndition` doesn't seem valid for the `btzen.set_trigger` function and thus I propose to change it to `condition` here. Please note this is just a visual observation and it is NOT tested at all.